Allow arbitrary json data for datasources This will allow us to close a bunch of issues in the provider Instead of supporting attributes one by one, users can just set a map The current JSONData objects will still be supported using the `Map()` function
